+++ examples/XMLReader/sitemap.php
@@ -38,3 +38,17 @@ if ($reader->read('s:url')) {
   } while ($reader->next('s:url'));
 }
 
+$reader = new FluentDOM\XMLReader();
+$reader->open('data://text/plain;base64,'.base64_encode($xml));
+$reader->registerNamespace('s', 'http://www.sitemaps.org/schemas/sitemap/0.9');
+$reader->registerNamespace('v', 'http://www.google.com/schemas/sitemap-video/1.1');
+
+foreach (new FluentDOM\XMLReader\SiblingIterator($reader, 's:url') as $url) {
+  var_dump(
+    [
+      $url('string(v:video/v:title)'),
+      $url('string(s:loc)')
+    ]
+  );
+}
+